Split async data types into separate modules
