+++ package.json
@@ -36,6 +36,12 @@
 	"exports": {
 		"./_*": null,
 		"./annotation.js": "./dist/annotation.js",
+		"./async.js": "./dist/async.js",
+		"./async/annotation.js": "./dist/async/annotation.js",
+		"./async/either.js": "./dist/async/either.js",
+		"./async/ior.js": "./dist/async/ior.js",
+		"./async/maybe.js": "./dist/async/maybe.js",
+		"./async/validation.js": "./dist/async/validation.js",
 		"./builder.js": "./dist/builder.js",
 		"./cmb.js": "./dist/cmb.js",
 		"./cmp.js": "./dist/cmp.js",
